Big change to iteration logging

+++ src/io/HMpsFF.cpp
@@ -207,7 +207,7 @@ HMpsFF::parsekey HMpsFF::parseDefault(std::ifstream& file) {
     HMpsFF::parsekey key = checkFirstWord(strline, s, e, word);
     if (key == HMpsFF::parsekey::NAME) {
       // Save name of the MPS file
-      if (e < strline.length()) {
+      if (e < (int)strline.length()) {
         mpsName = first_word(strline, e);
       }
       return HMpsFF::parsekey::NONE;
